BACKWARDS INCOMPATIBLE CHANGE
This is a substantial data model change where we move all errors to entity groups by project.  It should result in better performance and better ability to filter by project, for example, for stats.

Since this is a large change, we've renamed the model class such that incompatible old model instances will no longer be shown in the UI.  If anyone needs to migrate data, let us know and we can write a migration mapper.

Also: sets the stage for multiple levels of errors + deletes an unused function.